Various material related shader updates

+++ code/render/raytracing/raytracingcontext.h
@@ -43,7 +43,7 @@ public:
     static void Create(const RaytracingSetupSettings& settings);
 
     /// Setup an entity for ray tracing, assumes model context registration
-    static void Setup(const Graphics::GraphicsEntityId id, CoreGraphics::BlasInstanceFlags flags, uint mask);
+    static void Setup(const Graphics::GraphicsEntityId id, CoreGraphics::BlasInstanceFlags flags, uchar mask);
 
     /// Build top level acceleration
     static void ReconstructTopLevelAcceleration(const Graphics::FrameContext& ctx);
